added patch request in [notesId].js
