+++ pages/api/assignedGroup.js
@@ -23,7 +23,7 @@ export default async function assignedGroupingsHandler(req, res) {
     }
   } else if (req.method === "POST") {
     try {
-      const { group_assignment_id, student_id, group_id } = req.body;
+      const { student_id, group_id } = req.body;
       console.log(req.body);
       const assignGroup = await sql`
                  INSERT INTO assigned_student_groupings (student_id, group_id ) VALUES (${student_id}, ${group_id}) RETURNING *`;
